Backend/requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14,<4.0
django-cors-headers>=4.3,<5.0
bcrypt>=4.1,<5.0
PyJWT>=2.8,<3.0
python-dotenv>=1.0,<2.0
